requirements.txt
django>=5.2.4
djangorestframework>=3.16.0
drf-yasg>=1.21.7 